create common base for tests
