+++ tests/navigation_tests.py
@@ -1,24 +1,9 @@
-import time
-import subprocess
-import threading 
-import sys
-
 import unittest
-import subprocess
-import os
 
+from .SimulationTestBase import SimulationTestBase
 from ..clients.python_client.CompleteClient import CompleteClient
 
-class NavigationTests(unittest.TestCase):
-
-    def setUp(self):
-        self.sim = subprocess.Popen([os.environ["GODOT_PATH"], "--main-pack", " Simulation-Factory-Godot/simu/simulation.pck",
-         "--scenario", os.environ["GITHUB_WORKSPACE"] + "/simu/scenarios/new_scenario.json",
-         "--environment", os.environ["GITHUB_WORKSPACE"] + "/simu/environments/new_environment.json"])
-
-    def tearDown(self):
-        self.sim.kill()
-        self.sim.wait()
+class NavigationTests(SimulationTestBase):
 
     def test_navigate_to(self):
 
@@ -57,3 +42,5 @@ class NavigationTests(unittest.TestCase):
         finally:
             client.kill()
             
+if __name__ == '__main__':
+    unittest.main()
